Fixes
Gradient norm ended up being 0, so I added a bunch of noise to make it not zero.
This essentially could also mean that P and Q are identical, so I added some noise to P if that was the case.
It still seems to have the same problem. I want to properly fix this or maybe even start again from scratch instead of using a lot of patchwork.